# swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/user/addUser": {
      "post": {
        "description": "创建新用户",
        "operationId": "UserController_addUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "用户信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "添加用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/user/removeUser/{id}": {
      "delete": {
        "description": "根据ID删除用户",
        "operationId": "UserController_removeUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/user/updateUser/{id}": {
      "patch": {
        "description": "根据ID更新用户信息",
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "用户信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/user/login": {
      "post": {
        "description": "用户登录验证",
        "operationId": "UserController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "登录信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "用户登录",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/user/findUser/{id}": {
      "get": {
        "description": "根据ID查找用户信息",
        "operationId": "UserController_findUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "查找用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/user/getRole/{id}": {
      "get": {
        "description": "根据ID获取角色权限",
        "operationId": "UserController_getRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取角色权限",
        "tags": [
          "用户管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "nestjs Demo",
    "description": "nestjs 项目",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "开发环境"
    },
    {
      "url": "http://127.0.0.1:3001/api/v1",
      "description": "API v1 版本"
    }
  ],
  "components": {
    "schemas": {
      "GetUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "testuser"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123"
          },
          "roles": {
            "type": "object",
            "description": "角色列表"
          },
          "gender": {
            "type": "string",
            "description": "性别",
            "example": "male"
          },
          "profile": {
            "type": "object",
            "description": "用户资料"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "testuser"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "hashedPassword"
          }
        },
        "required": [
          "id",
          "username",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "testuser"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
    }
  }
}